docs : re-organize files
