+++ client/src/index.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import "./index.css";
-// import MapTest from './pages/map-test/map-test';
-import MemoryRoad from "./pages/MemoryRoad";
+import MapTest from './pages/map-test/map-test';
+//import MemoryRoad from "./pages/MemoryRoad";
 
 // redux
 import { Provider } from "react-redux";
@@ -10,7 +10,7 @@ import store from "./redux/store/index";
 
 ReactDOM.render(
   <Provider store={store}>
-    <MemoryRoad />
+    <MapTest />
   </Provider>,
   document.getElementById("root")
 );
